requirements.txt
streamlit
openai
pandas
matplotlib
Pillow
tenacity
python-dotenv
numpy
faiss-cpu
google-cloud-aiplatform
python-dotenv